# apiDocumentation/swagger.yaml
openapi: 3.0.0
info:
  description: This is the ISE-Learning REST API
  version: 0.0.0
  title: ISE-Learning REST API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/v0/users':
    get:
      tags:
        - Users
      summary: Gets all users information
      operationId: getUsers
      description: |
        Gets all users information.
      responses:
        '200':
          description: Retrieves all users
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/user'
    post:
      tags:
        - Users
      summary: Creates an user
      operationId: createUser
      description: Creates an user based on payload content
      requestBody:
            required: true
            description: User's information
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/user'
      responses:
        '201':
          description: User created successfuly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
  '/v0/users/{userId}':  
    get:
      tags:
        - Users
      summary: Retreives an user
      operationId: getUser
      description: |
        Retrive an user's information.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User's unique id.
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/user'
    delete:
      tags:
        - Users
      summary: Deletes an user
      operationId: deleteUser
      description: |
        Delete an user.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User's unique id.
      responses:
        '204':
          description: Successful deleted user.
          content:
            application/json:
              schema:
                type: object   
...
  '/v0/challenges/{challengeId}':
    get:
      tags:
        - Challenges
      summary: Retreives a challenge
      operationId: getChallenge
      description: |
        Retrieves every challenge the user has read access to.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      responses:
        '200':
          description: Challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: object
                    $ref: '#/components/schemas/challenge'
    put:
      tags:
        - Challenges
      summary: Updates a challenge
      operationId: updateChallenge
      description: |
        Updates a challenge based on payload content and the id on the uri.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      requestBody:
        description: Challenge to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/challengeNoId'
      responses:
        '200':
          description: Successful updated challenge.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challenge'
    delete:
      tags:
        - Challenges
      summary: Deletes a challenge
      operationId: deleteChallenge
      description: |
        Delete a challenge based on the id on the uri.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
      responses:
        '204':
          description: Successful deleted challenge.
          content:
            application/json:
              schema:
                type: object
  '/v0/challenges/users/{userId}':
    get:
      tags:
        - Challenges
      summary: Retreives all challenges of a given user
      operationId: getUserChallenges
      description: |
        Retrieves every challenge of a given user.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User unique id.
        - in: query
          name: privacy
          schema:
            type: string
            enum:
              - private
              - public
          required: false
          description: >-
            Optional filtering condition determining whether public or private
            challenges will be returned.
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          required: false
          description: Optional parameter filtering challenges by tags.
      responses:
        '200':
          description: User challenge list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    type: array
                    description: List of challenges from a given user.
                    items:
                      $ref: '#/components/schemas/challenge'
...
  '/v0/challenges/{challengeId}/answers/{userId}':
    get:
      tags:
        - Challenges
      summary: Retreives a challenge answer
      operationId: getChallengeAnswer
      description: |
        Retrieves every challenge answer of the user for a specific challenge.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User unique id.
      responses:
        '200':
          description: Challenge answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challengeAnswer'
  '/v0/challengeAnswers':
    post:
      tags:
        - Challenge Answer
      summary: Creates a challenge answer
      operationId: createsChallengeAnswer
      description: |
        Creates a challenge answer.
      requestBody:
        description: Challenge answer to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/challengeAnswerNoID'
      responses:
        '201':
          description: Successfully created challenge answer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challengeAnswer'
  '/v0/challengeAnswers/{challengeAnswerId}':
    put:
      tags:
        - Challenge Answer
      summary: Updates a challenge answer
      operationId: updateChallengeAnswer
      description: |
        Updates a challenge answer.
      parameters:
        - in: path
          name: challengeAnswerId
          schema:
            type: integer
          required: true
          description: Challenge answer unique id.
      requestBody:
        description: Challenge to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/challengeAnswerNoID'
      responses:
        '200':
          description: Successful updated challenge answer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challengeAnswer'
    delete:
      tags:
        - Challenge Answer
      summary: Deletes a challenge answer
      operationId: deleteChallengeAnswer
      description: |
        Delete a challenge answer of the used for a specific challenge.
      parameters:
        - in: path
          name: challengeAnswerId
          schema:
            type: integer
          required: true
          description: Challenge answer unique id.
      responses:
        '204':
          description: Successful deleted challenge answer.
          content:
            application/json:
              schema:
                type: object
...
  '/v0/questionnaires/users/{userId}':
    get:
      tags:
        - Questionnaires
      summary: Retrieves all questionnaires of specified user
      operationId: getAllUserQuestionnaires
      description: |
        Retrieves a collection of questionnaires by its creator unique identifer.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User unique identifer.
      responses:
        '200':
          description: User's questionnaire collection.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
...
components:
  schemas:
...
    user:
      type: object
      properties:
        userId: 
          type: integer
          example: 523
          description: Unique id which uniquely identifies the user
        username:
          type: string
          example: "userOne"
          description: Unique string used to identify an user
        email:
          type: string
          example: "userOne@example.com"
          description: Email address which can be used to contact user
        name:
          type: string
          example: "John Doe"
          description: User's given name
...
    challenge:
      type: object
      required:
        - challengeID
        - challengeText
        - isPrivate
        - solutions
      properties:
        challengeID:
          type: integer
          example: 548
          description: Unique identifier of challenge.
        challengeText:
          type: string
          example: 'Implement Hello world, by printing Hello world into the console.'
          description: Test describing the challenge and its goal.
        isPrivate:
          type: boolean
          example: true
          description: >-
            Boolean value indicating whether the challenge can only be seen by
            its creator
        solutions:
          type: array
          description: List of challenges.
          items:
            $ref: '#/components/schemas/solution'
    challengeNoId:
      type: object
      required:
        - challengeText
        - isPrivate
        - solutions
      properties:
        challengeText:
          type: string
          example: 'Implement Hello world, by printing Hello world into the console.'
          description: Test describing the challenge and its goal.
        isPrivate:
          type: boolean
          example: true
          description: >-
            Boolean value indicating whether the challenge can only be seen by
            its creator
        solutions:
          type: array
          description: List of challenges.
          items:
            $ref: '#/components/schemas/solution'
...
    solution:
      type: object
      required:
        - solutionId
        - codeLanguage
        - challengeCode
        - solutionCode
        - unitTests
      properties:
        solutionId:
          type: integer
          example: 79884
          description: Unique id to which uniquely identifies the solution.
        codeLanguage:
          type: string
          example: java
          description: Text specifying in which language the solution is.
        challengeCode:
          type: string
          example: 'public class code { public static void main(String[] args) { }}'
          description: Default code of the code presented to user.
        solutionCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests to which the the user submitted will be tested against.
...
    challengeAnswer:
      type: object
      required:
        - challengeAnswerId
        - challengeId
        - userID
        - codeLanguage
        - answerCode
        - unitTests
        - isCorrect
      properties:
        challengeAnswerId:
          type: integer
          example: 6546
          description: Unique id to which uniquely identify the challenge answer.
        challengeId:
          type: integer
          example: 55874
          description: Unique challenge id to which the answer belongs to.
        userID:
          type: integer
          example: 54954
          description: Unique user id of user to which the answer belongs to.
        codeLanguage:
          type: string
          example: Java
          description: Text identifying the programming language of the answer.
        answerCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge submitted.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests developed by the user.
        isCorrect:
          type: boolean
          example: true
          description: Boolean indicating if this answer is correct or not.
    challengeAnswerNoID:
      type: object
      required:
        - challengeId
        - userID
        - codeLanguage
        - answerCode
        - unitTests
        - isCorrect
      properties:
        challengeId:
          type: integer
          example: 55874
          description: Unique challenge id to which the aswer belongs to.
        userID:
          type: integer
          example: 54954
          description: Unique user id of user to which the aswer belongs to.
        codeLanguage:
          type: string
          example: Java
          description: Text identifying the programming language of the answer.
        answerCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge submitted.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests developed by the user.
        isCorrect:
          type: boolean
          example: true
          description: Boolean indicating if this answer is correct or not.
...
    questionnaire:
      type: object
      required:
        - questionnaireId
        - description
        - timer
        - creatorId
        - creationDate
      properties:
        questionnaireId:
          type: integer
          example: 21313
          description: questionnaire unique identifier.
        description:
          type: string
          example: Questionnaire for course PS-1920SV
          description: Text describing the questionnaire.
        timer:
          type: number
          example: 360000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire.
            If value is NULL has no time limite.
        creatorId:
          type: integer
          example: 44598
          description: Unique user identifier of user who created the questionnaire.
        creationDate:
          type: string
          example: '2020/05/12'
          description: questionnaire's creation date in yyyy-MM-dd format
...